updated view.py

+++ studyBud/myproject/base/views.py
@@ -15,6 +15,8 @@ from .forms import RoomForm
 # ]
 
 def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect("home")
     if request.method == "POST":
         username = request.POST.get("username")
         password = request.POST.get("password")
@@ -73,6 +75,7 @@ def updateRoom(request, pk):
 
     if request.user != room.host:
         return HttpResponse("You are not allowed here")
+
     if request.method == "POST" :
        form = RoomForm(request.POST, instance=room) 
        if form.is_valid():
@@ -85,6 +88,10 @@ def updateRoom(request, pk):
 @login_required(login_url="login")
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here")
+        
     if request.method == "POST":
         room.delete()
         return redirect("home")
